+++ packages/image-steam-bench/src/files.js
@@ -1,0 +1,10 @@
+const byIndex = ['12mp.jpeg', '18mp.jpeg', '24mp.jpeg'];
+const byKeys = byIndex.reduce((state, fn, idx) => {
+  state[fn] = idx;
+  return state;
+}, {});
+
+module.exports = {
+  byKeys,
+  byIndex
+}