+++ packages/image-steam-bench/src/test/index.js
@@ -1,0 +1,1 @@
+module.exports = ['origin', 'optimized', 'cached', 'real-90', 'real-95'];