+++ packages/image-steam-bench/src/util/sleep.js
@@ -1,0 +1,5 @@
+module.exports = async (timeout = 10000) => {
+  return new Promise(resolve => {
+    setTimeout(resolve, timeout).unref();
+  });
+}